utils/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.0=pypi_0
annotated-types=0.7.0=pypi_0
asttokens=3.0.0=pypi_0
attrs=25.3.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.6.15=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.4.127=0
cuda-runtime=12.4.1=0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.12.1=pypi_0
cython=3.1.2=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
easydict=1.13=pypi_0
eval-type-backport=0.2.2=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.16.1=pypi_0
fonttools=4.57.0=pypi_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.3.0=pypi_0
giflib=5.2.2=h5eee18b_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.40.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.70.0=pypi_0
hf-xet=1.1.5=pypi_0
huggingface-hub=0.33.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipython=8.12.3=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=py38h06a4308_0
jpeg=9b=h024ee3a_2
jpeg4py=0.1.4=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.8.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.7=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.30=0
libnvfatbin=12.4.127=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.48.0=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.17.0=h9b100fa_0
llvm-openmp=14.0.6=h9e868ea_0
lmdb=1.6.2=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py38h06a4308_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
ninja=1.12.1=h06a4308_0
ninja-base=1.12.1=hdb19cb5_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.3.1=pypi_0
opencv-python=4.11.0.86=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.16=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=py38hace64e9_1
pip=24.2=py38h06a4308_0
pipreqs=0.5.0=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.3.6=pypi_0
prompt-toolkit=3.0.51=pypi_0
protobuf=5.29.5=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.1.4=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.4.1=py3.8_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pyyaml=6.0.2=py38h5eee18b_0
pyzmq=27.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.4=pypi_0
requests-oauthlib=2.0.0=pypi_0
rpds-py=0.20.1=pypi_0
rsa=4.9.1=pypi_0
safetensors=0.5.3=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=2.32.0=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=59.5.0=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
soupsieve=2.7=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.13.3=py38h06a4308_0
tb-nightly=2.14.0a20230808=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tikzplotlib=0.10.1=pypi_0
timm=0.5.4=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.14=h993c535_1
torchaudio=2.4.1=py38_cu124
torchtriton=3.0.0=py38
torchvision=0.20.0=py38_cu124
tornado=6.4.2=pypi_0
tqdm=4.66.5=py38h2f386ee_0
traitlets=5.14.3=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.2.3=py38h06a4308_0
visdom=0.2.4=pypi_0
wandb=0.20.1=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py38h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarg=0.1.9=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.4.9=haebb681_0
